encoding.rs and value.rs cleanup
