+++ crates/document/src/encoding.rs
@@ -491,147 +491,3 @@ pub fn canonical_cbor_key_order(a: &&str, b: &&str) -> std::cmp::Ordering {
         other => other,
     }
 }
-
-#[cfg(test)]
-mod json_to_normal_value_for_kind_tests {
-    use super::*;
-    use schema::{ScalarArrayKind, ScalarKind};
-    use serde_json::json;
-
-    #[test]
-    fn datetime_string_becomes_time_not_string() {
-        // The regression: a DateTime field MUST coerce to Time so reindexed
-        // index entries match freshly-written ones (encode_time, not encode_string).
-        let nv =
-            json_to_normal_value_for_kind(&json!("2026-05-29T13:06:28Z"), &ScalarKind::DateTime);
-        let expected = DateTime::parse_from_rfc3339("2026-05-29T13:06:28Z").unwrap();
-        assert_eq!(nv, Some(NormalValue::Time(expected)));
-    }
-
-    #[test]
-    fn datetime_unix_timestamp_becomes_time() {
-        let nv = json_to_normal_value_for_kind(&json!(1_764_421_588_i64), &ScalarKind::DateTime);
-        match nv {
-            Some(NormalValue::Time(_)) => {}
-            other => panic!("expected Time, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn scalar_kinds_coerce_as_expected() {
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!(42), &ScalarKind::Int),
-            Some(NormalValue::Int(42))
-        );
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!(1.5), &ScalarKind::Float64),
-            Some(NormalValue::Float64(1.5))
-        );
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!(1.5), &ScalarKind::Float32),
-            Some(NormalValue::Float32(1.5))
-        );
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!(true), &ScalarKind::Bool),
-            Some(NormalValue::Bool(true))
-        );
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!("hi"), &ScalarKind::String),
-            Some(NormalValue::String("hi".to_string()))
-        );
-    }
-
-    #[test]
-    fn non_nillable_arrays_reject_null_elements() {
-        let cases = [
-            (json!([true, null]), ScalarArrayKind::BoolArray),
-            (json!([1, null]), ScalarArrayKind::IntArray),
-            (json!([1.5, null]), ScalarArrayKind::Float64Array),
-            (json!([1.5, null]), ScalarArrayKind::Float32Array),
-            (json!(["value", null]), ScalarArrayKind::StringArray),
-            (
-                json!(["2026-05-29T13:06:28Z", null]),
-                ScalarArrayKind::DateTimeArray,
-            ),
-        ];
-
-        for (value, kind) in cases {
-            assert_eq!(
-                json_to_normal_value_for_array_kind(&value, &kind),
-                None,
-                "{kind:?} must not invent a default for a null element"
-            );
-        }
-    }
-
-    #[test]
-    fn nillable_array_preserves_null_elements() {
-        assert_eq!(
-            json_to_normal_value_for_array_kind(
-                &json!([true, null]),
-                &ScalarArrayKind::NillableBoolArray,
-            ),
-            Some(NormalValue::NillableBoolElementArray(vec![
-                Some(true),
-                None,
-            ]))
-        );
-        assert_eq!(
-            json_to_normal_value_for_array_kind(
-                &json!(["2026-05-29T13:06:28Z", null]),
-                &ScalarArrayKind::NillableDateTimeArray,
-            ),
-            Some(NormalValue::NillableTimeElementArray(vec![
-                Some(DateTime::parse_from_rfc3339("2026-05-29T13:06:28Z").unwrap()),
-                None,
-            ]))
-        );
-    }
-
-    #[test]
-    fn coerce_stored_datetime_string_to_time() {
-        // The #72 repair: a DateTime read back from CBOR storage is a String;
-        // it must coerce to Time so the index encoder matches live-write entries.
-        let v = NormalValue::String("2026-05-29T13:06:28Z".to_string());
-        let got = coerce_stored_value_for_kind(v, &ScalarKind::DateTime);
-        let expected = DateTime::parse_from_rfc3339("2026-05-29T13:06:28Z").unwrap();
-        assert_eq!(got, NormalValue::Time(expected));
-    }
-
-    #[test]
-    fn coerce_stored_value_is_noop_for_non_datetime_and_correct_values() {
-        // A genuine String field is untouched (even if it looks date-like).
-        let s = NormalValue::String("2026-05-29T13:06:28Z".to_string());
-        assert_eq!(
-            coerce_stored_value_for_kind(s.clone(), &ScalarKind::String),
-            s
-        );
-        // An already-correct Time is untouched.
-        let t = NormalValue::Time(DateTime::parse_from_rfc3339("2026-05-29T13:06:28Z").unwrap());
-        assert_eq!(
-            coerce_stored_value_for_kind(t.clone(), &ScalarKind::DateTime),
-            t
-        );
-        // A non-RFC3339 string for a DateTime field is left as-is (no panic).
-        let bad = NormalValue::String("not-a-date".to_string());
-        assert_eq!(
-            coerce_stored_value_for_kind(bad.clone(), &ScalarKind::DateTime),
-            bad
-        );
-    }
-
-    #[test]
-    fn unrepresentable_value_returns_none() {
-        // A non-RFC3339 string for a DateTime field cannot be coerced; the caller
-        // decides whether that's an error (create) or a fallback (reindex).
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!("not-a-date"), &ScalarKind::DateTime),
-            None
-        );
-        // Json/None kinds are handled by callers, not here.
-        assert_eq!(
-            json_to_normal_value_for_kind(&json!({"a": 1}), &ScalarKind::Json),
-            None
-        );
-    }
-}
